# specs/001-buildcare-ai-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Buildcare AI API
  version: 0.1.0
  description: |
    Buildcare AI 백엔드 REST 계약. 근거: specs/001-buildcare-ai-platform/spec.md (FR-xxx),
    Design/SD_01 (P0~P8, G0~G10), Design/SD_02 (S1~S8B), Design/buildcare_ddl.sql.

    공통 규칙
    - 인증: httpOnly 쿠키 `bc_session`(JWT). 비회원은 `bc_device` 쿠키. 상태 변경 요청은 `X-CSRF-Token` 필요.
    - 게이트 미통과는 항상 `GateBlock` 본문으로 응답한다(FR-004). 문구는 gate_def.block_message 와 같다.
      HTTP 상태: G0 미로그인 401 · G0 권한 없음 403 · G1 402 · G2 402 · G3 422 · G5 422 · G4 502 · G6 409 · G7 200(blocked 필드) · G8 409 · G9 409 · G10 409
    - 결제·분석 요청·기록 저장·요청 확정은 `Idempotency-Key` 헤더를 받는다(중복 제출 방지).
    - 개인정보는 마스킹 필드만 반환한다(FR-090).
servers:
  - url: http://localhost:3000
    description: 로컬 개발 백엔드
security:
  - sessionCookie: []
...
paths:
...
  /api/cases/{caseId}:
    get:
      tags: [case]
      summary: 하자 건 요약 C6 (사진·설명·AI 결과·현장 기록·판정)
      parameters: [{ $ref: '#/components/parameters/CaseId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseSummary' }
        '403': { $ref: '#/components/responses/Gate' }
  /api/cases/{caseId}/progress:
    get:
      tags: [case]
      summary: 하자 건 진행 레일 C1 (v_case_progress)
      parameters: [{ $ref: '#/components/parameters/CaseId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseRail' }
...
  /api/records:
    post:
      tags: [record]
      summary: 현장 기록 시작 (draft). resultId 를 주면 사진·원인·대응방안 프리필 (FR-031)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [buildingId]
              properties:
                buildingId: { type: integer }
                resultId: { type: integer, nullable: true }
                scheduleId: { type: integer, nullable: true, description: 배정 점검에서 시작한 경우 }
      responses:
        '201':
          description: draft 생성
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InspectionRecord' }
        '403': { $ref: '#/components/responses/Gate' }
  /api/records/{recordId}:
    get:
      tags: [record]
      summary: 기록 조회
      parameters: [{ $ref: '#/components/parameters/RecordId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InspectionRecord' }
    patch:
      tags: [record]
      summary: draft 항목 수정 (임시 저장)
      parameters: [{ $ref: '#/components/parameters/RecordId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecordFields' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InspectionRecord' }
        '409':
          description: 저장된 기록은 repairStatus 외 수정 불가
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/records/{recordId}/photos:
    post:
      tags: [record]
      summary: 현장 사진 추가 (열린 추가 자료 요청 자동 해소 — FR-044)
      parameters: [{ $ref: '#/components/parameters/RecordId' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photos]
              properties:
                photos:
                  type: array
                  items: { type: string, format: binary }
      responses:
        '201':
          description: 저장된 사진 id 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  photoIds: { type: array, items: { type: integer } }
                  dataRequestResolved: { type: boolean }
        '502':
          description: 사진 저장 실패 — 텍스트는 draft 로 남음 (FR-035)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/records/{recordId}/save:
    post:
      tags: [record]
      summary: 기록 저장 (draft → saved). G5 필수 항목·사진 1장 이상
      parameters:
        - $ref: '#/components/parameters/RecordId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 저장 완료. mismatch 면 verificationItemId 포함 (FR-036)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/InspectionRecord'
                  - type: object
                    properties:
                      verificationItemId: { type: integer, nullable: true }
                      completedScheduleId: { type: integer, nullable: true }
        '422':
          description: G5 — missingFields 포함
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GateBlock' }
  /api/records/{recordId}/repair-status:
    patch:
      tags: [record]
      summary: 저장된 기록의 보수 결과 갱신 (FR-034, record_repair_log)
      parameters: [{ $ref: '#/components/parameters/RecordId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repairStatus]
              properties:
                repairStatus: { type: string, enum: [completed, pending] }
                repairMethod: { type: string, maxLength: 200 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InspectionRecord' }

  # ---------------------------------------------------------------- verification (S4)
...
  /api/buildings/{buildingId}/schedules:
    get:
      tags: [schedule]
      summary: 정기점검 일정과 표시 상태 (v_schedule_status)
      parameters: [{ $ref: '#/components/parameters/BuildingId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Schedule' }
    post:
      tags: [schedule]
      summary: 일정 생성·배정 (배정 통지 — FR-071)
      parameters: [{ $ref: '#/components/parameters/BuildingId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemText, assigneeId, dueDate]
              properties:
                itemText: { type: string, maxLength: 200 }
                cycleCode: { type: string, enum: [monthly, quarterly, semiannual, annual], nullable: true }
                assigneeId: { type: integer }
                dueDate: { type: string, format: date }
                basisPattern:
                  description: 반복 하자 근거로 항목 추가 시 (FR-070)
                  type: object
                  properties:
                    locationText: { type: string }
                    defectType: { type: string }
      responses:
        '201':
          description: 생성
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
  /api/schedules/{scheduleId}:
    patch:
      tags: [schedule]
      summary: 일정 수정 (scheduled 만)
      parameters: [{ $ref: '#/components/parameters/ScheduleId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                itemText: { type: string }
                cycleCode: { type: string, nullable: true }
                assigneeId: { type: integer }
                dueDate: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
  /api/schedules/{scheduleId}/cancel:
    post:
      tags: [schedule]
      summary: 일정 취소 — 이후 알림 없음 (FR-074)
      parameters: [{ $ref: '#/components/parameters/ScheduleId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
  /api/schedules/{scheduleId}/complete:
    post:
      tags: [schedule]
      summary: 같은 건물의 저장 기록을 연결해 완료 (FR-073). 기록에 위험 표시가 있으면 위험 통지 생성 (FR-076)
      parameters: [{ $ref: '#/components/parameters/ScheduleId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recordId]
              properties:
                recordId: { type: integer }
      responses:
        '200':
          description: 완료
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Schedule'
                  - type: object
                    properties:
                      riskNoticeId: { type: integer, nullable: true }
        '422':
          description: 다른 건물·저장 안 된 기록·이미 다른 일정의 근거
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /api/records/linkable-results:
    get:
      tags: [record]
      summary: 기록에 연결 가능한 분석 결과(비회원 미연결 결과 제외 — FR-120b)
      parameters: [{ name: buildingId, in: query, required: true, schema: { type: integer } }]
      responses: { '200': { description: OK } }
...
  /api/admin/permission-requests:
    get:
      tags: [admin]
      parameters: [{ name: status, in: query, schema: { type: string, enum: [pending, all] } }]
      responses: { '200': { description: OK } }
...
  /api/admin/rbac/roles/{role}/permissions:
    put:
      tags: [admin]
      summary: 역할 권한 변경(운영자) — 잠금 권한은 운영자에게서 제거 불가
      parameters: [{ name: role, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { permissions: { type: array, items: { type: string } } } } } } }
      responses: { '200': { description: OK }, '409': { description: 잠금 권한 } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bc_session
    deviceCookie:
      type: apiKey
      in: cookie
      name: bc_device
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 64 }
...
    CaseId: { name: caseId, in: path, required: true, schema: { type: integer } }
    BuildingId: { name: buildingId, in: path, required: true, schema: { type: integer } }
    RecordId: { name: recordId, in: path, required: true, schema: { type: integer } }
...
    ScheduleId: { name: scheduleId, in: path, required: true, schema: { type: integer } }
...
  responses:
    Gate:
      description: 게이트 미통과 — C2 차단 블록으로 표시
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GateBlock' }
...
  schemas:
    Problem:
      type: object
      required: [title]
      properties:
        title: { type: string }
        detail: { type: string }
        retryable: { type: boolean }
    GateBlock:
      type: object
      description: FR-004~FR-007. message 는 gate_def.block_message 와 동일
      required: [gate, message, releaseParty, actions]
      properties:
        gate: { type: string, enum: [G0, G1, G2, G3, G4, G5, G6, G7, G8, G9, G10] }
        message: { type: string }
        reason: { type: string, description: 구체 사유 (예 '사진이 흐립니다') }
        releaseParty: { type: string }
        selfRelease: { type: boolean, description: false 면 «요청 보내기» 행동만 (FR-006) }
        actions:
          type: array
          items:
            type: object
            properties:
              id: { type: string, example: retry-upload }
              label: { type: string, example: 사진 다시 올리기 }
              href: { type: string }
        missingFields:
          type: array
          items: { type: string, enum: [location, defectType, repairStatus, aiMatch, photo] }
        gateEventId: { type: integer }
...
    Code:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
...
    Notice:
      type: object
      properties:
        noticeId: { type: integer }
        kind: { type: string, enum: [analysis, priority, share] }
        body: { type: string }
        effectiveFrom: { type: string, format: date-time }
...
    AnalysisResult:
      type: object
      description: notice 는 항상 존재 (FR-014). recommendation 은 caution/danger 일 때 (FR-015)
      required: [resultId, riskLevel, causes, actions, notice]
      properties:
        resultId: { type: integer }
        riskLevel: { type: string, enum: [normal, caution, danger] }
        aiConfidence: { type: number, nullable: true }
        causes:
          type: array
          items:
            type: object
            properties:
              rank: { type: integer }
              text: { type: string }
        actions:
          type: array
          items:
            type: object
            properties:
              seq: { type: integer }
              text: { type: string }
        notice: { $ref: '#/components/schemas/Notice' }
        recommendation:
          type: object
          nullable: true
          properties:
            text: { type: string }
            canRequestExpert: { type: boolean, description: 건물관리자만 true (FR-121) }
            riskNoticeId: { type: integer, nullable: true }
        completedAt: { type: string, format: date-time }
    CaseSummary:
      type: object
      properties:
        caseId: { type: integer }
        caseNo: { type: string }
        buildingId: { type: integer, nullable: true }
        photos:
          type: array
          items:
            type: object
            properties:
              kind: { type: string, enum: [analysis, record] }
              photoId: { type: integer }
              alt: { type: string }
        description: { type: string, nullable: true }
        analysis:
          oneOf:
            - $ref: '#/components/schemas/AnalysisResult'
            - type: 'null'
        record:
          oneOf:
            - $ref: '#/components/schemas/InspectionRecord'
            - type: 'null'
        currentVerdict:
          oneOf:
            - $ref: '#/components/schemas/Verdict'
            - type: 'null'
    CaseRail:
      type: object
      description: 단계 상태 done · current · blocked:Gn · pending · n/a (n/a 는 '해당 없음' 글자 표기 — FR-008)
      properties:
        caseNo: { type: string }
        analysis: { type: string }
        record: { type: string }
        verify: { type: string }
        action: { type: string }
...
    RecordFields:
      type: object
      properties:
        locationText: { type: string, maxLength: 200 }
        defectType: { type: string }
        repairStatus: { type: string, enum: [completed, pending] }
        repairMethod: { type: string, maxLength: 200 }
        inspectionNote: { type: string, maxLength: 1000 }
        aiMatch: { type: string, enum: [match, mismatch, none] }
        riskFlag: { type: boolean, description: UC6 E3 점검 결과 위험 (ext) }
    InspectionRecord:
      allOf:
        - $ref: '#/components/schemas/RecordFields'
        - type: object
          properties:
            recordId: { type: integer }
            caseId: { type: integer }
            caseNo: { type: string }
            buildingId: { type: integer }
            buildingType: { $ref: '#/components/schemas/Code' }
            resultId: { type: integer, nullable: true }
            status: { type: string, enum: [draft, saved] }
            photoIds: { type: array, items: { type: integer } }
            savedAt: { type: string, format: date-time, nullable: true }
...
    Verdict:
      type: object
      properties:
        verdictId: { type: integer }
        itemId: { type: integer }
        versionNo: { type: integer }
        expertNameMasked: { type: string }
        verdict: { type: string, enum: [match, mismatch] }
        opinion: { type: string, nullable: true }
        diffNote: { type: string, nullable: true }
        riskHigh: { type: boolean }
        decidedAt: { type: string, format: date-time }
...
    Schedule:
      type: object
      properties:
        scheduleId: { type: integer }
        buildingId: { type: integer }
        buildingName: { type: string }
        itemText: { type: string }
        cycleCode: { type: string, nullable: true }
        assigneeId: { type: integer }
        assigneeNameMasked: { type: string }
        dueDate: { type: string, format: date }
        displayStatus: { type: string, enum: [scheduled, due, overdue, completed, cancelled] }
        completedRecordId: { type: integer, nullable: true }
...